requirements.txt
fastapi==0.115.11
pydantic==2.10.6
mysql-connector-python==9.0.0
streamlit==1.43.0
requests==2.32.3
pandas==2.2.3